+++ scitos_mira/launch/mira_launch.py
@@ -14,6 +14,13 @@ def generate_launch_description():
         'config',
         'mira_config.yaml'
         )
+    
+    # Read the robot XML configuration file.
+    scitos_config = os.path.join(
+        get_package_share_directory('scitos_mira'),
+        'resources/CLARC/',
+        'SCITOSDriver.xml'
+        )
 
     # Prepare the scitos_mira node.
     scitos_mira_node = Node(
@@ -21,7 +28,7 @@ def generate_launch_description():
             namespace = '',
             executable = 'scitos_mira',
             name = 'scitos_mira',
-            parameters = [config],
+            parameters = [config, {'scitos_config': scitos_config}],
             emulate_tty = True
         )